note this strange check against the 'ca-certs-version' file
what is even in this file?  what are we checking for exactly?

this `fgrep` check could still be wrong, but i am guessing at least
that "$OUR_CERTS_VERSION" should be a fixed string rather than a regex,
in particular for dots (eg, 1.2.3)

there are several other tests in this script with the form
`cat ca-certs-version | grep ...`, which to me are suspect